Move auth types and schemas to dedicated file
Auth-related types and Zod schemas were relocated from zuoraSchemas.ts to a new types/auth.ts file for better separation of concerns and maintainability.
