+++ modules/zuora/src/zuoraSchemas.ts
@@ -2,22 +2,6 @@ import { BillingPeriodValues } from '@modules/billingPeriod';
 import { z } from 'zod';
 import { zuoraResponseSchema } from './types';
 
-// --------------- Auth ---------------
-export type OAuthClientCredentials = z.infer<
-	typeof oAuthClientCredentialsSchema
->;
-export const oAuthClientCredentialsSchema = z.object({
-	clientId: z.string(),
-	clientSecret: z.string(),
-});
-
-export type ZuoraBearerToken = z.infer<typeof zuoraBearerTokenSchema>;
-
-export const zuoraBearerTokenSchema = z.object({
-	access_token: z.string(),
-	expires_in: z.number(),
-});
-
 // --------------- Subscription ---------------
 export const zuoraSubscriptionSchema = z.object({
 	id: z.string(),
